# docs/openapi.v1.yaml
openapi: 3.0.3
info:
  title: MADAVOLA API
  version: v1
  description: API contract v1 (spec-first). Certains endpoints sont planifiés.
servers:
  - url: /api/v1
...
paths:
...
  /territories/import:
    post:
      tags: [territories]
      summary: Import territory Excel
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                version_tag:
                  type: string
                file:
                  type: string
                  format: binary
              required: [version_tag, file]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TerritoryImportResult"
...
  /territories/versions/{version_tag}:
    get:
      tags: [territories]
      summary: Get territory version
      parameters:
        - in: path
          name: version_tag
          required: true
          schema: { type: string }
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: #/components/schemas/TerritoryVersion
...
  /territories/districts:
    get:
      tags: [territories]
      summary: List districts by region
      parameters:
        - in: query
          name: region_code
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/District"
  /territories/communes:
    get:
      tags: [territories]
      summary: List communes by district
      parameters:
        - in: query
          name: district_code
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Commune"
  /territories/fokontany:
    get:
      tags: [territories]
      summary: List fokontany by commune
      parameters:
        - in: query
          name: commune_code
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Fokontany"
...
  /transactions/{transaction_id}/payments:
    get:
      tags: [transactions]
      summary: List payments for transaction
      parameters:
        - in: path
          name: transaction_id
          required: true
          schema: { type: integer }
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: #/components/schemas/TransactionPayment
...
  /payments:
    get:
      tags: [payments]
      summary: List payments
      parameters:
        - in: query
          name: payer_actor_id
          schema: { type: integer }
        - in: query
          name: payee_actor_id
          schema: { type: integer }
        - in: query
          name: status
          schema: { type: string }
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: #/components/schemas/PaymentRequest
...
  /payments/webhooks/{provider}:
    post:
      tags: [payments]
      summary: Provider webhook
      x-implementation-status: planned
      parameters:
        - in: path
          name: provider
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
...
  /payments/{payment_id}:
    get:
      tags: [payments]
      summary: Get payment
      parameters:
        - in: path
          name: payment_id
          required: true
          schema: { type: integer }
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: #/components/schemas/PaymentRequest
  /payments/status/{external_ref}:
    get:
      tags: [payments]
      summary: Payment status
      x-implementation-status: planned
      parameters:
        - in: path
          name: external_ref
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
...
components:
  schemas:
...
    District:
      type: object
      properties:
        code: { type: string }
        name: { type: string }
        region_code: { type: string }
    Commune:
      type: object
      properties:
        code: { type: string }
        name: { type: string }
        district_code: { type: string }
        commune_mobile_money_msisdn: { type: string, nullable: true }
    Fokontany:
      type: object
      properties:
        code: { type: string, nullable: true }
        name: { type: string }
        commune_code: { type: string }
    TerritoryVersion:
      type: object
      properties:
        version_tag: { type: string }
        source_filename: { type: string }
        checksum_sha256: { type: string }
        status: { type: string }
        imported_at: { type: string, format: date-time }
        activated_at: { type: string, format: date-time, nullable: true }
    TerritoryImportResult:
      type: object
      properties:
        version_tag: { type: string }
        regions: { type: integer }
        districts: { type: integer }
        communes: { type: integer }
        fokontany: { type: integer }
...
    TransactionPayment:
      type: object
      properties:
        payment_request_id: { type: integer }
        payment_id: { type: integer }
        status: { type: string }
        external_ref: { type: string }
...
    PaymentRequest:
      type: object
      properties:
        id: { type: integer }
        provider_id: { type: integer }
        payer_actor_id: { type: integer }
        payee_actor_id: { type: integer }
        fee_id: { type: integer, nullable: true }
        transaction_id: { type: integer, nullable: true }
        amount: { type: number }
        currency: { type: string }
        status: { type: string }
        external_ref: { type: string }
...