# services/api/app/payments/router.py
import hashlib
import json
from pathlib import Path
from datetime import datetime, timezone
from uuid import uuid4

from fastapi import APIRouter, Depends, Request
from sqlalchemy.orm import Session

from app.common.errors import bad_request
from app.core.config import settings
from app.db import get_db
from app.audit.logger import write_audit
from app.auth.dependencies import get_current_actor
from app.models.actor import Actor, ActorRole
from app.models.fee import Fee
from app.models.invoice import Invoice
from app.models.document import Document
from app.models.payment import Payment, PaymentProvider, PaymentRequest, WebhookInbox
from app.models.transaction import TradeTransaction
from app.payments.schemas import (
    PaymentInitiate,
    PaymentInitiateResponse,
    PaymentRequestOut,
    WebhookPayload,
)

router = APIRouter(prefix=f"{settings.api_prefix}/payments", tags=["payments"])

@router.post("/initiate", response_model=PaymentInitiateResponse, status_code=201)
def initiate_payment(
    payload: PaymentInitiate,
    db: Session = Depends(get_db),
    current_actor=Depends(get_current_actor),
):
    provider = db.query(PaymentProvider).filter_by(code=payload.provider_code).first()
    if not provider or not provider.enabled:
        raise bad_request("provider_indisponible")

    payer = db.query(Actor).filter_by(id=payload.payer_actor_id).first()
    payee = db.query(Actor).filter_by(id=payload.payee_actor_id).first()
    if not payer or not payee:
        raise bad_request("acteur_invalide")
    if not _is_admin(db, current_actor.id) and payload.payer_actor_id != current_actor.id:
        raise bad_request("acces_refuse")

    fee = None
    if payload.fee_id:
        fee = db.query(Fee).filter_by(id=payload.fee_id).first()
        if not fee or fee.status != "pending":
            raise bad_request("frais_invalide")
        if fee.actor_id != payload.payer_actor_id:
            raise bad_request("frais_invalide")

    transaction = None
    if payload.transaction_id:
        transaction = (
            db.query(TradeTransaction).filter_by(id=payload.transaction_id).first()
        )
        if not transaction or transaction.status != "pending_payment":
            raise bad_request("transaction_invalide")
        if transaction.buyer_actor_id != payload.payer_actor_id:
            raise bad_request("transaction_invalide")

    if payload.external_ref:
        existing = db.query(PaymentRequest).filter_by(external_ref=payload.external_ref).first()
        if existing:
            payment = (
                db.query(Payment).filter_by(payment_request_id=existing.id).first()
            )
            return PaymentInitiateResponse(
                payment_request_id=existing.id,
                payment_id=payment.id if payment else 0,
                status=existing.status,
                external_ref=existing.external_ref,
            )

    external_ref = payload.external_ref or uuid4().hex
    request = PaymentRequest(
        provider_id=provider.id,
        payer_actor_id=payload.payer_actor_id,
        payee_actor_id=payload.payee_actor_id,
        fee_id=fee.id if fee else None,
        transaction_id=transaction.id if transaction else None,
        amount=payload.amount,
        currency=payload.currency,
        status="pending",
        external_ref=external_ref,
        idempotency_key=payload.idempotency_key,
    )
    db.add(request)
    db.flush()
    payment = Payment(payment_request_id=request.id, status="pending")
    db.add(payment)
    db.commit()
    write_audit(
        db,
        actor_id=payload.payer_actor_id,
        action="payment_initiated",
        entity_type="payment_request",
        entity_id=str(request.id),
        meta={"external_ref": request.external_ref, "amount": str(request.amount)},
    )
    db.commit()

    return PaymentInitiateResponse(
        payment_request_id=request.id,
        payment_id=payment.id,
        status=request.status,
        external_ref=request.external_ref,
    )


@router.post("/webhooks/{provider_code}")
async def webhook(provider_code: str, request: Request, db: Session = Depends(get_db)):
    provider = db.query(PaymentProvider).filter_by(code=provider_code).first()
    if not provider:
        raise bad_request("provider_inconnu")

    if settings.webhook_shared_secret:
        secret = request.headers.get("X-Webhook-Secret")
        if secret != settings.webhook_shared_secret:
            raise bad_request("webhook_non_autorise")

    if settings.webhook_ip_allowlist:
        allowed = {ip.strip() for ip in settings.webhook_ip_allowlist.split(",") if ip.strip()}
        client_ip = request.client.host if request.client else ""
        if allowed and client_ip not in allowed:
            raise bad_request("webhook_non_autorise")

    payload = await request.json()
    try:
        parsed = WebhookPayload(**payload)
    except Exception:
        raise bad_request("payload_invalide")

    payload_hash = hashlib.sha256(json.dumps(payload, sort_keys=True).encode()).hexdigest()
    existing = (
        db.query(WebhookInbox)
        .filter_by(provider_id=provider.id, external_ref=parsed.external_ref)
        .first()
    )
    if existing:
        return {"status": "ok", "idempotent": True}

    inbox = WebhookInbox(
        provider_id=provider.id,
        external_ref=parsed.external_ref,
        payload_hash=payload_hash,
        status="received",
    )
    db.add(inbox)

    payment_request = db.query(PaymentRequest).filter_by(external_ref=parsed.external_ref).first()
    if payment_request:
        payment_request.status = parsed.status
        payment = db.query(Payment).filter_by(payment_request_id=payment_request.id).first()
        if payment:
            payment.status = parsed.status
            if parsed.status == "success":
                payment.confirmed_at = datetime.now(timezone.utc)
        if parsed.status == "success" and payment_request.fee_id:
            fee = db.query(Fee).filter_by(id=payment_request.fee_id).first()
            if fee and fee.status != "paid":
                fee.status = "paid"
                fee.paid_at = datetime.now(timezone.utc)
                actor = db.query(Actor).filter_by(id=fee.actor_id).first()
                if actor and actor.status == "pending":
                    actor.status = "active"
                write_audit(
                    db,
                    actor_id=fee.actor_id,
                    action="fee_paid",
                    entity_type="fee",
                    entity_id=str(fee.id),
                    meta={"payment_request_id": payment_request.id},
                )
        if parsed.status == "success" and payment_request.transaction_id:
            transaction = (
                db.query(TradeTransaction)
                .filter_by(id=payment_request.transaction_id)
                .first()
            )
            if transaction and transaction.status != "paid":
                transaction.status = "paid"
                invoice_number = f"INV-{transaction.id:08d}"
                invoice = Invoice(
                    invoice_number=invoice_number,
                    transaction_id=transaction.id,
                    seller_actor_id=transaction.seller_actor_id,
                    buyer_actor_id=transaction.buyer_actor_id,
                    total_amount=transaction.total_amount,
                    status="issued",
                )
                db.add(invoice)
                storage_dir = Path(settings.document_storage_dir)
                storage_dir.mkdir(parents=True, exist_ok=True)
                invoice_payload = {
                    "invoice_number": invoice_number,
                    "transaction_id": transaction.id,
                    "seller_actor_id": transaction.seller_actor_id,
                    "buyer_actor_id": transaction.buyer_actor_id,
                    "total_amount": str(transaction.total_amount),
                    "currency": transaction.currency,
                }
                content = json.dumps(invoice_payload, ensure_ascii=True).encode()
                sha256 = hashlib.sha256(content).hexdigest()
                filename = f"{invoice_number}.json"
                storage_path = storage_dir / filename
                storage_path.write_bytes(content)
                db.add(
                    Document(
                        doc_type="invoice",
                        owner_actor_id=transaction.seller_actor_id,
                        related_entity_type="invoice",
                        related_entity_id=invoice_number,
                        storage_path=str(storage_path),
                        original_filename=filename,
                        sha256=sha256,
                    )
                )
                write_audit(
                    db,
                    actor_id=transaction.buyer_actor_id,
                    action="invoice_issued",
                    entity_type="invoice",
                    entity_id=invoice_number,
                    meta={"transaction_id": transaction.id},
                )
        if parsed.status == "success":
            write_audit(
                db,
                actor_id=payment_request.payer_actor_id,
                action="payment_success",
                entity_type="payment_request",
                entity_id=str(payment_request.id),
                meta={"external_ref": payment_request.external_ref},
            )

    db.commit()
    return {"status": "ok", "idempotent": False}


@router.get("", response_model=list[PaymentRequestOut])
def list_payments(
    payer_actor_id: int | None = None,
    payee_actor_id: int | None = None,
    status: str | None = None,
    db: Session = Depends(get_db),
    current_actor=Depends(get_current_actor),
):
    query = db.query(PaymentRequest)
    if not _is_admin(db, current_actor.id):
        query = query.filter(
            (PaymentRequest.payer_actor_id == current_actor.id)
            | (PaymentRequest.payee_actor_id == current_actor.id)
        )
        if payer_actor_id and payer_actor_id != current_actor.id:
            return []
        if payee_actor_id and payee_actor_id != current_actor.id:
            return []
    if payer_actor_id:
        query = query.filter(PaymentRequest.payer_actor_id == payer_actor_id)
    if payee_actor_id:
        query = query.filter(PaymentRequest.payee_actor_id == payee_actor_id)
    if status:
        query = query.filter(PaymentRequest.status == status)
    payments = query.order_by(PaymentRequest.created_at.desc()).all()
    return [
        PaymentRequestOut(
            id=p.id,
            provider_id=p.provider_id,
            payer_actor_id=p.payer_actor_id,
            payee_actor_id=p.payee_actor_id,
            fee_id=p.fee_id,
            transaction_id=p.transaction_id,
            amount=float(p.amount),
            currency=p.currency,
            status=p.status,
            external_ref=p.external_ref,
        )
        for p in payments
    ]


@router.get("/{payment_id}", response_model=PaymentRequestOut)
def get_payment(
    payment_id: int,
    db: Session = Depends(get_db),
    current_actor=Depends(get_current_actor),
):
    payment = db.query(PaymentRequest).filter_by(id=payment_id).first()
    if not payment:
        raise bad_request("paiement_introuvable")
    if not _is_admin(db, current_actor.id):
        if current_actor.id not in (payment.payer_actor_id, payment.payee_actor_id):
            raise bad_request("acces_refuse")
    return PaymentRequestOut(
        id=payment.id,
        provider_id=payment.provider_id,
        payer_actor_id=payment.payer_actor_id,
        payee_actor_id=payment.payee_actor_id,
        fee_id=payment.fee_id,
        transaction_id=payment.transaction_id,
        amount=float(payment.amount),
        currency=payment.currency,
        status=payment.status,
        external_ref=payment.external_ref,
    )


@router.get("/status/{external_ref}", response_model=PaymentRequestOut)
def get_payment_status(
    external_ref: str,
    db: Session = Depends(get_db),
    current_actor=Depends(get_current_actor),
):
    payment = db.query(PaymentRequest).filter_by(external_ref=external_ref).first()
    if not payment:
        raise bad_request("paiement_introuvable")
    if not _is_admin(db, current_actor.id):
        if current_actor.id not in (payment.payer_actor_id, payment.payee_actor_id):
            raise bad_request("acces_refuse")
    return PaymentRequestOut(
        id=payment.id,
        provider_id=payment.provider_id,
        payer_actor_id=payment.payer_actor_id,
        payee_actor_id=payment.payee_actor_id,
        fee_id=payment.fee_id,
        transaction_id=payment.transaction_id,
        amount=float(payment.amount),
        currency=payment.currency,
        status=payment.status,
        external_ref=payment.external_ref,
    )


def _is_admin(db: Session, actor_id: int) -> bool:
    return (
        db.query(ActorRole)
        .filter(ActorRole.actor_id == actor_id, ActorRole.role.in_(["admin", "dirigeant"]))
        .first()
        is not None
    )